+++ tests/JsonStringTest.php
@@ -39,7 +39,7 @@ class JsonStringTest extends TestCase
     function testRetrievingOfStringValue(): void
     {
         $str = new JsonString('Hello');
-        self::assertEquals('Hello', $str->getString());
+        self::assertEquals('Hello', $str->stringValue());
     }
 
     function testEquality(): void